+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## [Unreleased]
 
+## [0.1.99] - 2022-10-21
+
+- Shuffle deterministically tracks for non admin users on evolution detail view
+- Show which rating value was selected
+- Improve rating validation
+- Open youtube link in new tab
+
 ## [0.1.98] - 2022-10-20
 
 - Improved user rating based fitness adjustment